# app/server-ns/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/login": {
      "post": {
        "operationId": "MainController_login",
        "summary": "用户登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
    "/logout": {
      "post": {
        "operationId": "MainController_logout",
        "summary": "退出登录",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
    "/register": {
      "post": {
        "operationId": "MainController_register",
        "summary": "用户注册",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "根目录"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rovella-Admin",
    "description": "Rovella-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "userName",
          "password",
          "uuid"
        ]
      },
...
    }
  }
}